Move ValueToJSONType function to jsonnet pkg
