+++ internal/pkg/encoding/jsonnet/context.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/google/go-jsonnet"
 	"github.com/google/go-jsonnet/ast"
-	"github.com/spf13/cast"
 
 	"github.com/keboola/keboola-as-code/internal/pkg/utils/errors"
 )
@@ -29,46 +28,6 @@ type (
 type NativeFunction = jsonnet.NativeFunction
 
 type NotifierFactory func(ctx context.Context) jsonnet.Notifier
-
-// ValueToLiteral converts Go value to jsonnet.Ast literal.
-func ValueToLiteral(v interface{}) ast.Node {
-	if v == nil {
-		return &ast.LiteralNull{}
-	}
-
-	switch v := v.(type) {
-	case bool:
-		return &ast.LiteralBoolean{Value: v}
-	case int:
-		return &ast.LiteralNumber{OriginalString: cast.ToString(v)}
-	case int32:
-		return &ast.LiteralNumber{OriginalString: cast.ToString(v)}
-	case int64:
-		return &ast.LiteralNumber{OriginalString: cast.ToString(v)}
-	case float32:
-		return &ast.LiteralNumber{OriginalString: cast.ToString(v)}
-	case float64:
-		return &ast.LiteralNumber{OriginalString: cast.ToString(v)}
-	case []any:
-		elements := make([]ast.CommaSeparatedExpr, 0)
-		for _, aVal := range v {
-			elements = append(elements, ast.CommaSeparatedExpr{Expr: ValueToLiteral(aVal)})
-		}
-		return &ast.Array{Elements: elements}
-	case map[string]any:
-		fields := make(ast.DesugaredObjectFields, 0)
-		for mKey, mVal := range v {
-			fields = append(fields, ast.DesugaredObjectField{
-				Hide: ast.ObjectFieldInherit,
-				Name: ValueToLiteral(mKey),
-				Body: ValueToLiteral(mVal),
-			})
-		}
-		return &ast.DesugaredObject{Fields: fields}
-	default:
-		return &ast.LiteralString{Value: cast.ToString(v), Kind: ast.StringDouble}
-	}
-}
 
 func NewContext() *Context {
 	return &Context{
